# [Core] Do async init of xgrammar in the engine