@mgoin 2024-12-03
This makes quite a large difference on Llama-3.1-8B-Instruct on H100:

```
python benchmark_guided.py --model meta-llama/Llama-3.1-8B-Instruct --dataset xgrammar_bench --async-engine --output-len 512 --num-prompts 32 --enable-chunked-prefill --guided-decoding-ratio 1

# Main @ 2f2cdc745a7a569637c58cfd5f6789c1d0741c84
Throughput: 2.66 requests/s, 2189.98 total tokens/s, 1363.66 output tokens/s Correct rate is 93.75 % 
First token latency(msecs):
count      32.000000
mean     4076.612419
std       707.871948
min      3046.303682
25%      3654.368758
50%      4330.442839
75%      4826.109441
max      4862.703992
dtype: float64
Next token latency(msecs):
count    32.000000
mean     15.479507
std       1.348600
min      13.995258
25%      14.047387
50%      14.993054
75%      16.284354
max      17.443944
dtype: float64

# This PR
Throughput: 4.19 requests/s, 3442.13 total tokens/s, 2143.35 output tokens/s Correct rate is 93.75 % 
First token latency(msecs):
count      32.000000
mean     1104.019488
std        64.567038
min      1012.124414
25%      1066.139331
50%      1119.639536
75%      1171.954522
max      1184.854695
dtype: float64
Next token latency(msecs):
count    32.000000
mean     12.748215
std       0.093481
min      12.639430
25%      12.645627
50%      12.726120
75%      12.803509
max      12.883286
dtype: float64
```