@russellb 2024-12-04
While early results looked very promising, these changes no longer seemed to help once testing with other simpler optimizations in place. Using the online benchmark script from #10880, we can't justify the level of change introduced here. 

Part of the benefit here was in the cleanup and removing the need for serialization and passing around additional data. We can continue pursuing further cleanup and optimization in V1.